+++ UpSkill/Api/UpSkill.Api.csproj
@@ -26,4 +26,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Data\UpSkill.Data.csproj" />
+  </ItemGroup>
+
 </Project>
