+++ UpSkill/Api/WeatherForecast.cs
@@ -1,15 +1,15 @@
 using System;
 
-namespace UpSkillApi
+namespace UpSkill.Api
 {
-    public class WeatherForecast
-    {
-        public DateTime Date { get; set; }
+	public class WeatherForecast
+	{
+		public DateTime Date { get; set; }
 
-        public int TemperatureC { get; set; }
+		public int TemperatureC { get; set; }
 
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+		public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
 
-        public string Summary { get; set; }
-    }
+		public string Summary { get; set; }
+	}
 }
