+++ UpSkill/ClientSide/Program.cs
@@ -10,23 +10,23 @@ using System.Threading.Tasks;
 
 namespace UpSkill.ClientSide
 {
-    public class Program
-    {
-        public static async Task Main(string[] args)
-        {
-            var builder = WebAssemblyHostBuilder.CreateDefault(args);
-            builder.RootComponents.Add<App>("#app");
+	public class Program
+	{
+		public static async Task Main(string[] args)
+		{
+			var builder = WebAssemblyHostBuilder.CreateDefault(args);
+			builder.RootComponents.Add<App>("#app");
 
-            builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
+			builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
 
-            builder.Services.AddOidcAuthentication(options =>
-            {
-                // Configure your authentication provider options here.
-                // For more information, see https://aka.ms/blazor-standalone-auth
-                builder.Configuration.Bind("Local", options.ProviderOptions);
-            });
+			builder.Services.AddOidcAuthentication(options =>
+			{
+				// Configure your authentication provider options here.
+				// For more information, see https://aka.ms/blazor-standalone-auth
+				builder.Configuration.Bind("Local", options.ProviderOptions);
+			});
 
-            await builder.Build().RunAsync();
-        }
-    }
+			await builder.Build().RunAsync();
+		}
+	}
 }
