+++ UpSkill/Data/Common/Models/BaseDeletableModel.cs
@@ -1,11 +1,11 @@
-﻿namespace UpSkill.Infrastructure.Data.Common.Models
+﻿namespace UpSkill.Data.Common.Models
 {
-    using System;
+	using System;
 
-    public abstract class BaseDeletableModel<TKey> : BaseModel<TKey>, IDeletableEntity
-    {
-        public bool IsDeleted { get; set; }
+	public abstract class BaseDeletableModel<TKey> : BaseModel<TKey>, IDeletableEntity
+	{
+		public bool IsDeleted { get; set; }
 
-        public DateTime? DeletedOn { get; set; }
-    }
+		public DateTime? DeletedOn { get; set; }
+	}
 }
