+++ UpSkill/Data/Common/Models/BaseModel.cs
@@ -1,15 +1,15 @@
-﻿namespace UpSkill.Infrastructure.Data.Common.Models
+﻿namespace UpSkill.Data.Common.Models
 {
-    using System;
-    using System.ComponentModel.DataAnnotations;
+	using System;
+	using System.ComponentModel.DataAnnotations;
 
-    public abstract class BaseModel<TKey> : IAuditInfo
-    {
-        [Key]
-        public TKey Id { get; set; }
+	public abstract class BaseModel<TKey> : IAuditInfo
+	{
+		[Key]
+		public TKey Id { get; set; }
 
-        public DateTime CreatedOn { get; set; }
+		public DateTime CreatedOn { get; set; }
 
-        public DateTime? ModifiedOn { get; set; }
-    }
+		public DateTime? ModifiedOn { get; set; }
+	}
 }
