+++ UpSkill/Data/Common/Models/IAuditInfo.cs
@@ -1,11 +1,11 @@
-﻿namespace UpSkill.Infrastructure.Data.Common.Models
+﻿namespace UpSkill.Data.Common.Models
 {
-    using System;
+	using System;
 
-    public interface IAuditInfo
-    {
-        DateTime CreatedOn { get; set; }
+	public interface IAuditInfo
+	{
+		DateTime CreatedOn { get; set; }
 
-        DateTime? ModifiedOn { get; set; }
-    }
+		DateTime? ModifiedOn { get; set; }
+	}
 }
