+++ UpSkill/Data/Models/Grade.cs
@@ -1,22 +1,22 @@
-﻿namespace UpSkill.Infrastructure.Data.Models
+﻿namespace UpSkill.Data.Models
 {
-    using System;
-    using System.ComponentModel.DataAnnotations;
+	using System;
+	using System.ComponentModel.DataAnnotations;
 
-    public class Grade
-    {
-        [Key]
-        [Required]
-        public string Id { get; init; } = Guid.NewGuid().ToString();
+	public class Grade
+	{
+		[Key]
+		[Required]
+		public string Id { get; init; } = Guid.NewGuid().ToString();
 
-        public decimal Value { get; set; }
+		public decimal Value { get; set; }
 
-        [Required]
-        public string CourseId { get; set; }
-        public Course Course { get; set; }
+		[Required]
+		public string CourseId { get; set; }
+		public Course Course { get; set; }
 
-        [Required]
-        public string StudentId { get; set; }
-        public Employee Student { get; set; }
-    }
+		[Required]
+		public string StudentId { get; set; }
+		public Employee Student { get; set; }
+	}
 }
