+++ UpSkill/Data/Models/Owner.cs
@@ -1,14 +1,14 @@
-﻿namespace UpSkill.Infrastructure.Data.Models
+﻿namespace UpSkill.Data.Models
 {
-    using System.Collections.Generic;
+	using System.Collections.Generic;
 
-    public class Owner : ApplicationUser
-    {
-        public Owner()
-        {
-            this.Employees = new HashSet<Employee>();
-        }
+	public class Owner : ApplicationUser
+	{
+		public Owner()
+		{
+			this.Employees = new HashSet<Employee>();
+		}
 
-        public ICollection<Employee> Employees { get; set; }
-    }
+		public ICollection<Employee> Employees { get; set; }
+	}
 }
