+++ UpSkill/Data/Models/StudentCourse.cs
@@ -1,15 +1,15 @@
-﻿namespace UpSkill.Infrastructure.Data.Models
+﻿namespace UpSkill.Data.Models
 {
-    using System.ComponentModel.DataAnnotations;
+	using System.ComponentModel.DataAnnotations;
 
-    public class StudentCourse
-    {
-        [Required]
-        public string StudentId { get; set; }
-        public Employee Student { get; set; }
+	public class StudentCourse
+	{
+		[Required]
+		public string StudentId { get; set; }
+		public Employee Student { get; set; }
 
-        [Required]
-        public string CourseId { get; set; }
-        public Course Course { get; set; }
-    }
+		[Required]
+		public string CourseId { get; set; }
+		public Course Course { get; set; }
+	}
 }
