--- UpSkill/Services/UpSkill.Services.Data/Class1.cs
@@ -1,8 +1,0 @@
-﻿namespace UpSkill.Services.Data
-{
-    using System;
-
-    public class Class1
-    {
-    }
-}
